api/agent/requirements.txt
fastapi
fastapi[standard]
mangum
uvicorn
pydantic
strands-agents
openai
pymongo
httpx